Remove useless tests
